Video player Refactored
