+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AndroidUdemyCourse">
-
+        <activity android:name=".feature.media_player.activities.VideoPlayerActivity"></activity>
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
@@ -20,12 +20,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
         <activity
             android:name=".feature.atm_consultancy.AtmConsultancyActivity"
             android:label="@string/atm_title"
             android:theme="@style/Theme.AtmConsultancy" />
-
         <activity
             android:name=".feature.media_player.activities.MediaPlayerActivity"
             android:label="@string/media_player"
