+++ app/src/main/res/layout/fragment_video.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".feature.media_player.fragments.VideoPlayerFragment">
+    tools:context=".feature.media_player.fragments.VideoFragment">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/backgroundVideoImageView"
@@ -75,6 +75,4 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/ratingVideoRatingBar" />
 
-
 </androidx.constraintlayout.widget.ConstraintLayout>
-
