--- app/src/main/res/layout/fragment_video_host.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".feature.media_player.fragments.VideoHostFragment">
-
-
-     <VideoView
-         android:id="@+id/videoHostvideoView"
-         android:layout_width="0dp"
-         android:layout_height="0dp"
-         app:layout_constraintBottom_toBottomOf="parent"
-         app:layout_constraintEnd_toEndOf="parent"
-         app:layout_constraintStart_toStartOf="parent"
-         app:layout_constraintTop_toTopOf="parent"
-         tools:layout_editor_absoluteX="0dp"
-         tools:layout_editor_absoluteY="0dp" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
